Move everything that is exports except the module itself to wasi_types
